@GitHK review: doc and rename error_codes
