+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/task.py
@@ -23,7 +23,7 @@ from fastapi import FastAPI
 from models_library.projects_networks import DockerNetworkAlias
 from models_library.projects_nodes_io import NodeID
 from models_library.service_settings_labels import RestartPolicy
-from servicelib.errors import create_error_code
+from servicelib.error_codes import create_error_code
 
 from ....core.settings import (
     DynamicServicesSchedulerSettings,
